+++ api/__main__.py
@@ -1,6 +1,7 @@
-from . import app, routes, PORT
+from . import app, PORT
+from .v1 import routes as v1_routes
 from os import getenv
 
-routes.add_all(app)
+v1_routes.registry(app)
 
 app.run('0.0.0.0', PORT, debug=getenv("DEBUG") != None)
