+++ api/v1/extractors/home.py
@@ -1,8 +1,8 @@
 from bs4 import BeautifulSoup
 from bs4.element import Tag
-from .. import SITE
-from ..utils import http_get
 from ..types import HomeResult
+from ... import SITE
+from ...utils import http_get
 
 
 async def get_all(page: int = 1) -> list[HomeResult]:
