+++ api/v1/extractors/search.py
@@ -1,8 +1,8 @@
 from bs4 import BeautifulSoup
 from bs4.element import Tag
-from .. import SITE
-from ..utils import http_get
 from ..types import SearchResult
+from ... import SITE
+from ...utils import http_get
 
 
 async def search(query: str, page: int = 1) -> tuple[list[SearchResult] | None, bool | str]:
