+++ api/v1/types.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from .utils import Dataclass2Dict
+from ..utils import Dataclass2Dict
 
 
 @dataclass
